+++ engine/src/Engine/YStaticMesh.cpp
@@ -63,7 +63,7 @@ void YStaticMesh::Render(CameraBase* camera)
 	int triangle_total = 0;
 	for (auto& polygon_group : raw_meshes[0]->polygon_groups)
 	{
-		int triangle_count = polygon_group.polygons.size();
+		int triangle_count =(int) polygon_group.polygons.size();
 		dc->DrawIndexed(triangle_count * 3, triangle_total, 0);
 		triangle_total += triangle_count * 3;
 	}
